rename: 문제 관련 인터페이스 디렉토리 변경
